requirements.txt
dvc>=3.29.0
dvclive>=3.0.1
scikit-learn==1.3.2
pandas
numpy
plotly.express
ucimlrepo
pyyaml